requirements.txt
# Единый список зависимостей проекта (root + KinoServer + model).
# --- Общие утилиты ---
python-dotenv
requests

# --- Bot (Telegram) ---
aiogram
asyncpg
aiofiles

# --- Parser / Rating (PostgreSQL) ---
psycopg2-binary
deep-translator==1.11.4
googletrans

# --- Dataset tools ---
numpy
pandas

# --- Embeddings / ML ---
sentence-transformers
huggingface-hub
transformers==4.57.3
torch==2.9.1

# --- KinoServer (FastAPI) ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic-settings>=2.4.0
aiosqlite>=0.20.0

